Move utils to own folder
