+++ src/components/VideoProvider/useRoom/useRoom.tsx
@@ -1,6 +1,6 @@
 import { Callback } from '../../../types';
 import EventEmitter from 'events';
-import { isMobile } from '../../../utils';
+import { isMobile } from '../../../utils/utils';
 import Video, { ConnectOptions, LocalTrack, Room } from 'twilio-video';
 import { useCallback, useEffect, useRef, useState } from 'react';
